mbed_retarget: Make filehandles optional to reduce code size
The retarget code allocates an array of FileHandle* for console and file
handling (filehandles). A tiny target only needs a console (putc/getc).
There is no need for file handling.
The POSIX layer and the array of FileHandle* is not required for small
targets that only need a console ; this code should be optionally compiled
out if the configuration parameter `platform.stdio-console-only` is
selected instead of `platform-stdio-buffered-serial`.