CMake: Create an independent cmsis library
The rtos library (to be created) depends on cmsis. To avoid having
a circular dependency between the mbed-os library and the rtos
library, a separate cmsis library has been created so the rtos
library can link with it.
However, because the cmsis library in turn depends on Mbed target
specific implementation (which is part of the mbed-os library)
of CMSIS, there is a circular dependency
(mbed-os->rtos->cmsis->mbed-os).
Therefore, all the libraries have to be made static as per CMake rule
on circular dependency.
We can look into creating a separate library that contains all the
Mbed target specific source (`targets` library?) in the future.