CMake: Move connectivity targets creation to parent directory
This was done as some targets are used across different directories.

For example mbed-os-mbedtls is used both in connectivity/drivers
and connectivity/mbedtls. However, as the target was created in the latter,
it was not visible in the former. The creation of the target had
therefore to be in a parent directory of both directories.

Similar issue as the example above for ble, nfc, etc.